Refactor header guard package and handle trailing comments
